rm-flow-store: move billing code in AccountStore
