+++ bucket/ddc_bucket/account/mod.rs
@@ -1,2 +1,3 @@
 pub mod entity;
 pub mod store;
+pub mod messages;
