--- bucket/ddc_bucket/billing/mod.rs
@@ -1,1 +1,0 @@
-pub mod messages;
