--- bucket/ddc_bucket/flow/mod.rs
@@ -1,1 +1,0 @@
-pub mod entity;
