+++ bucket/lib.rs
@@ -15,11 +15,8 @@ pub mod ddc_bucket {
     use cluster::{entity::*, store::*};
     use deal::{entity::*, store::*};
     use Error::*;
-    use flow::entity::*;
-    use schedule::*;
     use vnode::{entity::*, store::*};
 
-    pub mod billing;
     pub mod account;
     pub mod flow;
     pub mod schedule;
